+++ router/hourRouter.py
@@ -1,27 +1,38 @@
 from fastapi import APIRouter, HTTPException,Depends
-from schema.hourSchema import RequestHour, Response
+from schema.hourSchema import RequestHour
 from service.hourService import HourService
-from config.config import sessionlocal
-from sqlalchemy.orm import Session
 from typing import Annotated
 from fastapi import status
 from sqlalchemy.orm.exc import NoResultFound
+from service.tokenHandler import TokenHandler
+from model.role import Role
+from model.user import User
+from service.userService import get_current_user
 
 router = APIRouter()
 
 
         
 dependency = Annotated[HourService,Depends()]
+user_dependency = Annotated[User,Depends(get_current_user)]
 
 # get all hours end point
 @router.get("/")
-async def get_all(hour_service: dependency):
+@TokenHandler.role_required([Role.ADMIN,Role.STUDENT,Role.TEACHER])
+async def get_all(user:user_dependency,hour_service: dependency):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,
+                                detail="Authentication failed")
     _hours = hour_service.get_hours()
     return _hours
      
 # get hour by id end point
 @router.get("/{hour_id}")
-async def get(hour_id: int,hour_service: dependency ):
+@TokenHandler.role_required([Role.ADMIN,Role.STUDENT,Role.TEACHER])
+async def get(user:user_dependency,hour_id: int,hour_service: dependency):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,
+                                detail="Authentication failed")
     try:
         _hour = hour_service.get_hour_by_id(hour_id=hour_id)
         return _hour
@@ -31,14 +42,22 @@ async def get(hour_id: int,hour_service: dependency ):
 
 # create hour end point
 @router.post("/create")
-async def create(hour: RequestHour, hour_service: dependency):
+@TokenHandler.role_required([Role.ADMIN,Role.TEACHER])
+async def create(user:user_dependency,hour: RequestHour, hour_service: dependency):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,
+                                detail="Authentication failed")
     _hour = hour_service.create_hour(hour=hour)
     return _hour
       
 
 # delete hour end point
 @router.delete("/delete/{hour_id}")
-async def delete(hour_id: int, hour_service: dependency):
+@TokenHandler.role_required([Role.ADMIN,Role.TEACHER])
+async def delete(user:user_dependency,hour_id: int, hour_service: dependency):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,
+                                detail="Authentication failed")
     try:
        _hour = hour_service.delete_hour(hour_id=hour_id)
        return _hour
@@ -48,7 +67,11 @@ async def delete(hour_id: int, hour_service: dependency):
 
 # update hour end point 
 @router.patch("/update/{hour_id}")
-async def update(hour_id: int, hour: RequestHour, hour_service: dependency):
+@TokenHandler.role_required([Role.ADMIN,Role.TEACHER])
+async def update(user:user_dependency,hour_id: int, hour: RequestHour, hour_service: dependency):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,
+                                detail="Authentication failed")
     try:
         _hour = hour_service.update_hour(
                             hour_id=hour_id,
